introducing the app factory pattern for better testability: the app is now created via the create_app() function in src/__init__.py. the errors subsystem was refactored in its own package for better maintainability und to prevent circular imports
