+++ api/Makefile
@@ -1,6 +1,7 @@
 SH := /bin/sh
 PY := $(shell which python3)
 MULTIPROC_DIR ?= $(shell pwd)/.multiproc_dir
+CONFFILE = devconf.py
 
 .PHONY: help
 ## help: prints this help message
@@ -11,6 +12,7 @@ help:
 .PHONY: test
 ## test: run unit tests
 test:
+	export CONFFILE=$(CONFFILE); \
 	cd src/ && $(PY) -m pytest -s -x
 
 .PHONY: run
@@ -20,16 +22,19 @@ run: run/dev
 .PHONY: run/dev
 ## run/dev: start via the Flask development server
 run/dev:
+	CONFFILE=$(CONFFILE) \
+	FLASK_APP=./src/run.py \
 	ENVIRONMENT=development \
 	FLASK_ENV=development \
-	FLASK_APP=./src/run.py \
-	FLASK_DEBUG=1 \
+	FLASK_DEBUG=True \
 		flask run
 
 .PHONY: run/prod
 ## run/prod: start via gunicorn
 run/prod:
-	ENVIRONMENT=production ./gunicorn.sh $(MULTIPROC_DIR)
+	CONFFILE=$(CONFFILE) \
+	ENVIRONMENT=production \
+	./gunicorn.sh $(MULTIPROC_DIR)
 
 .PHONY: clean
 ## clean: cleans .multiproc_dir
