+++ api/src/devconf.py
@@ -1,0 +1,6 @@
+"""
+Configuration file for the development target
+"""
+
+ERROR_404_HELP = False
+OPENAPI_SPEC = 'openapi/spec.yml'
